:feet: Move generate imports/files methods to base_components.
